Refactor model into separate module
