+++ app.py
@@ -1,18 +1,10 @@
 from flask import Flask, render_template, request, redirect, url_for
-from flask_sqlalchemy import SQLAlchemy
-from datetime import datetime
+from models import db, MaintenanceLog
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///maintenance.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy(app)
-
-class MaintenanceLog(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    date = db.Column(db.Date, default=datetime.utcnow)
-    part = db.Column(db.String(100), nullable=False)
-    mileage = db.Column(db.Integer)
-    notes = db.Column(db.String(200))
+db.init_app(app)
 
 with app.app_context():
     db.create_all()
